golf_booking_backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
dj-rest-auth==7.0.1
django-allauth==65.11.2
